# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Service {
  id: string;
  name: string;
  category: string;
  description: string;
  durationMinutes: number;
  priceCents: number;
}

export interface Stylist {
  id: string;
  name: string;
  role: string;
  bio: string;
  imageUrl: string;
  specialties: string[];
}

export interface AvailabilitySlot {
  startsAt: string;
  endsAt: string;
  available: boolean;
}

export interface AvailabilityResponse {
  date: string;
  slots: AvailabilitySlot[];
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  notes?: string | null;
  serviceId: string;
  stylistId: string;
  startsAt: string;
  endsAt: string;
  status: AppointmentStatus;
  createdAt: string;
  service?: Service;
  stylist?: Stylist;
}

export interface CreateAppointmentInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 3 */
  customerEmail: string;
  /** @minLength 3 */
  customerPhone: string;
  serviceId: string;
  stylistId: string;
  startsAt: string;
  notes?: string;
}

export type UpdateAppointmentStatusInputStatus =
  (typeof UpdateAppointmentStatusInputStatus)[keyof typeof UpdateAppointmentStatusInputStatus];

export const UpdateAppointmentStatusInputStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface UpdateAppointmentStatusInput {
  status: UpdateAppointmentStatusInputStatus;
}

export interface TopService {
  serviceId: string;
  name: string;
  bookings: number;
}

export interface DashboardSummary {
  appointmentsToday: number;
  appointmentsThisWeek: number;
  upcomingCount: number;
  revenueThisWeekCents: number;
  topService?: TopService | null;
}

export type GetAvailabilityParams = {
  serviceId: string;
  stylistId: string;
  /**
   * ISO date (YYYY-MM-DD) in salon local time
   */
  date: string;
};

export type ListAppointmentsParams = {
  from?: string;
  to?: string;
  status?: ListAppointmentsStatus;
  stylistId?: string;
};

export type ListAppointmentsStatus =
  (typeof ListAppointmentsStatus)[keyof typeof ListAppointmentsStatus];

export const ListAppointmentsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export type GetUpcomingAppointmentsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};