# lib/api-zod/src/generated/types/availabilityResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilitySlot } from "./availabilitySlot";

export interface AvailabilityResponse {
  date: string;
  slots: AvailabilitySlot[];
}