# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  startsAt: string;
  endsAt: string;
  available: boolean;
}