# lib/api-zod/src/generated/types/createAppointmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 3 */
  customerEmail: string;
  /** @minLength 3 */
  customerPhone: string;
  serviceId: string;
  stylistId: string;
  startsAt: string;
  notes?: string;
}