# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TopService } from "./topService";

export interface DashboardSummary {
  appointmentsToday: number;
  appointmentsThisWeek: number;
  upcomingCount: number;
  revenueThisWeekCents: number;
  topService?: TopService | null;
}