# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  name: string;
  category: string;
  description: string;
  durationMinutes: number;
  priceCents: number;
}