# lib/api-zod/src/generated/types/updateAppointmentStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAppointmentStatusInputStatus } from "./updateAppointmentStatusInputStatus";

export interface UpdateAppointmentStatusInput {
  status: UpdateAppointmentStatusInputStatus;
}